+++ Backend/src/routes/billing.ts
@@ -3,6 +3,7 @@ import {sql, stripe} from '../connection';
 import {Elysia, error} from 'elysia';
 import {jwt} from '@elysiajs/jwt';
 import {attempt} from '../utils';
+import {check} from '../checks';
 
 export default new Elysia({prefix: '/billing'})
   .use(jwt({name: 'jwt', secret: process.env.JWT_SECRET as string}))
@@ -34,7 +35,12 @@ export default new Elysia({prefix: '/billing'})
     return {received: true};
   })
   .post('/setup', async ({body}) => {
-    console.log(body);
+    const {email, payment, err} = check(body, ['email', 'payment']);
+    if (err) return error(400, err);
+
+    const customer = await stripe.customers.create({email, payment_method: payment});
+    if (!customer) return error(500, 'Failed to create customer');
+    await attempt(sql`UPDATE users SET stripe_customer = ${customer.id} WHERE email = ${email}`);
   })
   .get('/checkout', async ({user, query}) => {
     if (!user) return error(401, 'You are not logged in');
@@ -46,13 +52,17 @@ export default new Elysia({prefix: '/billing'})
     if (!type) return error(400, 'Missing or invalid query type. Try again');
     if (!pricingModal[type]) return error(400, 'Invalid query type. Try again');
 
+    const response = await attempt(sql`SELECT stripe_customer FROM users WHERE email = ${user.email}`);
+    const customer = response[0].stripe_customer || '';
+    const option = customer ? {customer} : {customer_email: user.email};
+
     const session = await stripe.checkout.sessions.create({
       ui_mode: 'custom' as 'embedded',
       return_url: `${origin}/dashboard`,
       mode: type === 'lifetime' ? 'payment' : 'subscription',
       line_items: [{price: pricingModal[type], quantity: 1}],
       payment_method_types: ['card'],
-      customer_email: user!.email,
+      ...option,
     });
 
     return {clientSecret: session.client_secret};