+++ Frontend/src/routes/purchase/+page.svelte
@@ -34,17 +34,43 @@
   async function handleCheckout() {
     stripe.initCheckout({clientSecret}).then((checkout) => {
       const paymentElement = checkout.createElement('payment', {layout: 'tabs'});
-      const button = document.getElementById('pay-button')!;
-      const errors = document.getElementById('confirm-errors')!;
+      checkout.changeAppearance({
+        theme: 'flat',
+        variables: {
+          colorPrimary: '#4338ca',
+          colorBackground: '#1e293b',
+          colorText: '#cbd5e1',
+          colorDanger: '#ef4444',
+          colorSuccess: '#22c55e',
+          colorWarning: '#eab308',
+          colorTextPlaceholder: '#475569',
+          fontFamily: 'Inter, sans-serif',
+          fontWeightNormal: '500',
+          spacingUnit: '4px',
+          borderRadius: '8px',
+          tabLogoColor: 'dark',
+        },
+        rules: {
+          '.Input': {
+            marginTop: '14px',
+            marginBottom: '8px',
+            border: '2px solid #1e293b',
+            backgroundColor: '#131c2e',
+          },
+          '.Input:focus': {
+            boxShadow: '0 1px 1px 0 rgba(0, 0, 0, 0.07), 0 0 0 2px #4f46e5',
+          },
+        },
+      });
 
-      paymentElement.mount('#payment-element');
-      button.addEventListener('click', () => {
-        errors.textContent = '';
+      paymentElement.mount('#payment');
+      document.querySelector('button[name="pay"]')!.addEventListener('click', async () => {
+        fetching.set(1);
+        await new Promise((resolve) => setTimeout(resolve, 650));
 
         checkout.confirm().then((result) => {
-          if (result.type === 'error') {
-            errors.textContent = result.error.message;
-          }
+          fetching.set(0);
+          if (result.type === 'error') notify(result.error.message, 'alert');
         });
       });
     });
@@ -69,7 +95,7 @@
   <form method="POST" use:enhance={handleSubmit}>
     <h1 class="text-6xl">Purchase a new identity</h1>
     <p>Purchase an affordable identity to protect your privacy and safeguard your data.</p>
-    <div class="flex justify-center p-8">
+    <div class="flex justify-center gap-8 p-8">
       {#each ['Monthly', 'Annually', 'Lifetime'] as model}
         <button class:select-model={model === $pricingModel.name} type="button" onclick={() => changeModel(model)}>
           {model}
@@ -80,9 +106,10 @@
     <LoadingButton className="w-fit px-16 py-6 text-2xl mt-4">Purchase</LoadingButton>
   </form>
   <Modal id={1}>
-    <div id="payment-element"></div>
-    <button id="pay-button">Pay</button>
-    <div id="confirm-errors"></div>
+    <div class="m-4 flex flex-col gap-8">
+      <div id="payment" class="sm:w-80"></div>
+      <LoadingButton name="pay" index={1}>Pay</LoadingButton>
+    </div>
   </Modal>
 </div>
 
@@ -96,6 +123,6 @@
   }
 
   .select-model {
-    @apply bg-gradient-to-br from-green-500 to-green-800 text-white;
+    @apply bg-gradient-to-br from-green-500 to-green-700 text-white shadow-green-900 hover:shadow-green-950;
   }
 </style>