+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.10] - 2025-1-14
+
+### Added
+
+- Add design to payment form using stripe element
+- Add customer check in purchase page
+- Add payment to signup
+
 ## [0.3.09] - 2025-1-10
 
 ### Added